refactor: Jdbc 템플릿 DIP
