+++ src/main/java/persistence/Application.java
@@ -2,9 +2,10 @@ package persistence;
 
 import database.DatabaseServer;
 import database.H2;
-import persistence.jdbc.JdbcTemplate;
+import jdbc.BaseJdbcTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import persistence.entity.JdbcTemplate;
 
 public class Application {
     private static final Logger logger = LoggerFactory.getLogger(Application.class);
@@ -15,7 +16,7 @@ public class Application {
             final DatabaseServer server = new H2();
             server.start();
 
-            final JdbcTemplate jdbcTemplate = new JdbcTemplate(server.getConnection());
+            final JdbcTemplate jdbcTemplate = new BaseJdbcTemplate(server.getConnection());
 
             server.stop();
         } catch (Exception e) {
