+++ src/main/java/repository/DDLRepository.java
@@ -1,6 +1,6 @@
 package repository;
 
-import persistence.jdbc.JdbcTemplate;
+import persistence.entity.JdbcTemplate;
 import persistence.sql.QueryGenerator;
 
 public class DDLRepository {
